Address eyrun's comments from kristin's PR
